Learning rate finder/requirement.txt
numpy
matplotlib
scipy
tqdm
pytorch